# Rewrite hamming_distance function to a 17x faster recursive version